[REFACTOR] Modularize components in Wallet page
The **Voting Stake** and **Participation Reward** sections of the wallet page have been moved to their own components.

When locking and unlocking DGD, we need to wait for the transactions to be confirmed by the info-server. Since there's no reliable way (yet) to fetch this data and update the values on the page immediately after the action, we should disable the buttons to prevent the user from locking/unlocking more DGD until the transactions have been confirmed.

**Test Plan: Claim DGX**

- Run `npm run teleport:main_phase` on the `dao-contracts` repo then restart the `info-server`.
- Load the `account4` wallet and go to the Wallet Page.
- Claim your DGX. After the transaction, the button should be disabled while the DGX value remains the same.
- Confirm that the data has been updated in the `info-server`.
- Go to another tab then return to the Profile page. You should have zero claimable DGX.

**Test Plan: Lock DGD**

- Load any wallet then go to the Wallet page.
- Lock some DGD. After the transaction, the button should be disabled while the stake value remains the same.
- Confirmt hat the data has bee updated in the `info-server`.
- Go to another tab then return to the Profile page. The stake should be updated and the button should be enabled again.

**Test Plan: Unlock DGD**

- Run `npm run teleport:locking_phase` on the `dao-contracts` repo then restart the `info-server`.
- Repeat the test for **Lock DGD** with the **Unlock DGD** button. The results should be the same.
